Move away from deprecated java converters
